fix: pages 배포 오류 관련 이미지 폴더이름 변경및, 경로 수정
